minor changes to memset/ffsll? and string.h
memset doesn't need cld because DF is clear on function entry because
the abi says so

ffsl and ffsll return int so they don't need cltq

remove variable names from string.h

+++ lib/string/ffsl.c
@@ -1,1 +1,1 @@
-long ffsl(long i) { return __builtin_ffsl(i); }
+int ffsl(long i) { return __builtin_ffsl(i); }
